requirements.txt
asgiref==3.2.10
astroid==2.4.2
colorama==0.4.3
Django==3.0.7
django-templated-mail==1.1.1
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
djoser==2.0.3
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
Pillow==7.2.0
psycopg2==2.8.5
PyJWT==1.7.1
pylint==2.5.3
python-dotenv==0.14.0
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
typed-ast==1.4.1
wrapt==1.12.1
